+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- routes can be excluded using regex patterns
+
 ## [1.1.0] - 2017-01-31
 ### Added
 - unresolved paths outputted to the user
 - user can include list of paths to be  checked apart from resolved ones
 
-## [Unreleased]
-- remove constructor from PHPUnit test
-- use data provider in Codeception so that each route is a new test